requirements.txt
python-dotenv==0.21.1
python-telegram-bot-raw==13.12
requests==2.28.2
urllib3==1.26.14